libfreeipmi: use byte-sized VLAN priority API
Match the public API and its implementation to the byte-sized parameter
accepted by the LAN command builder.

+++ libfreeipmi/api/ipmi-lan-cmds-api.c
@@ -1306,7 +1306,7 @@ ipmi_cmd_set_lan_configuration_parameters_vlan_id (ipmi_ctx_t ctx,
 int
 ipmi_cmd_set_lan_configuration_parameters_vlan_priority (ipmi_ctx_t ctx,
                                                          uint8_t channel_number,
-                                                         uint32_t vlan_priority,
+                                                         uint8_t vlan_priority,
                                                          fiid_obj_t obj_cmd_rs)
 {
   fiid_obj_t obj_cmd_rq = NULL;
